requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-rest-framework==0.1.0
djangorestframework==3.12.4
flower==1.0.0
gevent==21.8.0
googlemaps==4.5.3
greenlet==1.1.2
humanize==3.11.0
idna==3.2
kombu==5.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycparser==2.20
pytz==2021.1
redis==3.5.3
requests==2.26.0
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0

